+++ VKTS_PKG_Window/src/window/visual/fn_visual_xlib.cpp
@@ -521,7 +521,7 @@ VkBool32 VKTS_APIENTRY _visualDispatchMessages()
 
 int32_t VKTS_APIENTRY _visualGetNumberDisplays()
 {
-    return g_allDisplays.size();
+    return (int32_t)g_allDisplays.size();
 }
 
 INativeDisplayWP VKTS_APIENTRY _visualCreateDefaultDisplay()
@@ -856,7 +856,7 @@ INativeWindowWP VKTS_APIENTRY _visualCreateWindow(const INativeDisplayWP& displa
     defaultScreen = DefaultScreen(g_nativeDisplay);
 
     long visualMask = VisualScreenMask;
-    visualInfoTemplate.screen = defaultScreen;
+    visualInfoTemplate.screen = (int)defaultScreen;
 
     visualInfo = XGetVisualInfo(g_nativeDisplay, visualMask, &visualInfoTemplate, &numberOfVisuals);
